requirements.txt
# ============================================================
# TTRPG Narrator — Dependencias Python
# ============================================================

# GUI
dearpygui>=1.11.0

# HTTP (Ollama)
requests>=2.31.0

# PDF parsing
PyMuPDF>=1.24.0

# YAML
pyyaml>=6.0.1

# Vault (frontmatter en archivos Markdown)
python-frontmatter>=1.1.0

# Plantillas de prompts
jinja2>=3.1.3

# ── Futuro (comentado hasta Sprint 2) ─────────────────────
# Búsqueda semántica con embeddings
# chromadb>=0.5.0
# sentence-transformers>=2.7.0

# CLI mejorada
# rich>=13.7.0

# Tests
pytest>=8.1.0
